requirements.txt
Pillow==8.2.0
tensorflow==1.13.1
tensorflow-hub==0.7.0
nltk==3.6.2
Keras==2.2.4
djangorestframework==3.12.4
gunicorn==20.1.0
django-cors-headers==3.7.0
numpy==1.19.5
django==3.2
h5py==2.10.0